refactor: move and split write concern spec tests
