+++ test/integration/read-write-concern/read_write_concern.spec.test.js
@@ -1,0 +1,18 @@
+'use strict';
+
+const { TestRunnerContext, generateTopologyTests } = require('../../tools/spec-runner');
+const { loadSpecTests } = require('../../spec');
+
+describe('Read Write Concern spec tests', function () {
+  describe('operation spec tests', function () {
+    const testContext = new TestRunnerContext();
+    const testSuites = loadSpecTests('read-write-concern/operation');
+
+    after(() => testContext.teardown());
+    before(function () {
+      return testContext.setup(this.configuration);
+    });
+
+    generateTopologyTests(testSuites, testContext);
+  });
+});
